js/src/xpconnect/tests/chrome/test_evalInSandbox.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=533596
-->
<window title="Mozilla Bug 533596"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">

  <iframe type="content"
          src="http://example.org/tests/js/src/xpconnect/tests/mochitest/file_evalInSandbox.html"
          onload="go()"
          id="ifr">
  </iframe>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[
      const Cu = Components.utils;
      const Ci = Components.interfaces;
      const utils = window.QueryInterface(Ci.nsIInterfaceRequestor)
                          .getInterface(Ci.nsIDOMWindowUtils);

      function go() {
        var win = $('ifr').contentWindow;
        var sandbox = new Cu.Sandbox(win);
        is(utils.getClassName(sandbox),
           "Proxy",
           "sandbox was wrapped correctly");
        sandbox.__proto__ = new XPCNativeWrapper(win);

        is(utils.getClassName(Cu.evalInSandbox("this.document", sandbox)),
           "Proxy",
           "return value was rewrapped correctly");
        ok(Cu.evalInSandbox("('wrappedJSObject' in this.document);", sandbox),
           "wrappers inside eIS are XPCNativeWrappers");

        SimpleTest.finish();
      }

      SimpleTest.waitForExplicitFinish();
  ]]></script>
</window>
